Converted all values to float64.

+++ glmat33/arg.go
@@ -2,7 +2,7 @@ package glmat33
 
 import "math"
 
-func (v *Vec2) Arg() float32 {
-	return float32(math.Atan2(float64(v[1]), float64(v[0])))
+func (v *Vec2) Arg() float64 {
+	return math.Atan2(v[1], v[0])
 }
