+++ glmat33/normalse.go
@@ -4,12 +4,12 @@ import "math"
 
 // Normalise the Vec4 v.
 func (v *Vec4) Normalise() *Vec4 {
-	h := float32(math.Sqrt(float64(v[0]*v[0] + v[1]*v[1] + v[2]*v[2] + v[3]*v[3])))
+	h := math.Sqrt(v[0]*v[0] + v[1]*v[1] + v[2]*v[2] + v[3]*v[3])
 	if h == 0 {
-		v[0] = float32(math.NaN())
-		v[1] = float32(math.NaN())
-		v[2] = float32(math.NaN())
-		v[3] = float32(math.NaN())
+		v[0] = math.NaN()
+		v[1] = math.NaN()
+		v[2] = math.NaN()
+		v[3] = math.NaN()
 		return v
 	}
 
@@ -32,11 +32,11 @@ func (v Vec4) Normalised() (normalised Vec4) {
 
 // Normalise the Vec3 v.
 func (v *Vec3) Normalise() *Vec3 {
-	h := float32(math.Sqrt(float64(v[0]*v[0] + v[1]*v[1] + v[2]*v[2])))
+	h := math.Sqrt(v[0]*v[0] + v[1]*v[1] + v[2]*v[2])
 	if h == 0 {
-		v[0] = float32(math.NaN())
-		v[1] = float32(math.NaN())
-		v[2] = float32(math.NaN())
+		v[0] = math.NaN()
+		v[1] = math.NaN()
+		v[2] = math.NaN()
 		return v
 	}
 
@@ -58,10 +58,10 @@ func (v Vec3) Normalised() (normalised Vec3) {
 
 // Normalise the Vec2 v.
 func (v *Vec2) Normalise() *Vec2 {
-	h := float32(math.Sqrt(float64(v[0]*v[0] + v[1]*v[1])))
+	h := math.Sqrt(v[0]*v[0] + v[1]*v[1])
 	if h == 0 {
-		v[0] = float32(math.NaN())
-		v[1] = float32(math.NaN())
+		v[0] = math.NaN()
+		v[1] = math.NaN()
 		return v
 	}
